Allow text location to be customised
It's helpful, particularly for label printing purposes, to be able to change
where the text is displayed relative to the QR code location.  This required
a certain amount of surgery to the image generation, to ensure correct
sizing and alignment in all cases.  I have tested this code in all four
orientations, and including multi-line output, and it appears to produce visually
appealing results, however some corner cases may remain that I didn't think of.